[plugin sparkline] hook up addcol-sparkline
- on ColumnsSheet, 'addcol-sparkline' adds a sparkline from the selected rows (columns on source sheet)
- on regular Sheet, 'addcol-sparkline' adds a sparkline from all visible numeric rows (like plot-numerics)
