+++ plugins/sparkline.py
@@ -1,0 +1,35 @@
+from visidata import *
+
+
+option('disp_sparkline', '▁▂▃▄▅▆▇', 'characters to display sparkline')
+
+
+class SparklineColumn(Column):
+    def calcValue(self, r):
+        return sparkline(*tuple(c.getTypedValue(r) for c in self.source))
+
+
+    def sparkline(*values):
+        lines = options.disp_sparkline
+        values = [int(v) for v in values]
+        mx = max(values)
+        mn = min(values)
+        w = (mx - mn) / len(lines)
+        bounds = [(mn + w * i) for i in range(len(lines))]
+
+        output = ''
+        for val in values:
+            for b in bounds:
+                if mn == 0 and val == 0:
+                    output += ' '
+                    break
+                if val < b:
+                    output += lines[bounds.index(b) - 1]
+                    break
+            else:
+                output += max(lines)
+        return output
+
+
+ColumnsSheet.addCommand(None, 'addcol-sparkline', 'cursorRow.sheet.addColumn(SparklineColumn("sparkline", source=selectedRows))')
+Sheet.addCommand(None, 'addcol-sparkline', 'addColumn(SparklineColumn("sparkline", source=numericCols(nonKeyVisibleCols)))')
