+++ visidata/__init__.py
@@ -80,7 +80,6 @@ from .layout import *
 from .saving import *
 import visidata.sort
 
-from .sparkline import *
 from .vimkeys import *
 
 from .loaders.csv import *
